JobBoard/backend/requirements.txt
aiohttp==3.8.4
bcrypt==4.0.1
fastapi==0.103.2
httpx==0.25.0
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyMySQL==1.1.0
pytest==7.4.2
python-dotenv==1.0.0
PyYAML==6.0.1
SQLAlchemy==2.0.21
sqlparse==0.4.3
uvicorn==0.23.2